+++ Godeps/_workspace/src/github.com/fd/go-nat/README.md
@@ -1,0 +1,3 @@
+# go-nat
+
+[![GoDoc](https://godoc.org/github.com/fd/go-nat?status.svg)](https://godoc.org/github.com/fd/go-nat) [![status](https://sourcegraph.com/api/repos/github.com/fd/go-nat/.badges/status.png)](https://sourcegraph.com/github.com/fd/go-nat)
